添加前台权限内容的逻辑与现实

+++ resources/views/post/show.blade.php
@@ -5,18 +5,23 @@
         <div class="blog-post">
             <div style="display:inline-flex">
                 <h2 class="blog-post-title">{{$post->title}}</h2>
-                <a style="margin: auto" href="/posts/{{$post->id}}/edit">
-                    <span class="glyphicon glyphicon-pencil" aria-hidden="true"></span>
-                </a>
-                <a style="margin: auto" href="/posts/{{$post->id}}/delete">
-                    <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
-                </a>
+                @can('update',$post)
+                    <a style="margin: auto" href="/posts/{{$post->id}}/edit">
+                        <span class="glyphicon glyphicon-pencil" aria-hidden="true"></span>
+                    </a>
+                @endcan
+                @can('delete',$post)
+                    <a style="margin: auto" href="/posts/{{$post->id}}/delete">
+                        <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
+                    </a>
+                @endcan
             </div>
 
-            <p class="blog-post-meta">{{$post->created_at->toFormattedDateString()}} by <a href="#">{{\Auth::user()->name}}</a></p>
+            <p class="blog-post-meta">{{$post->created_at->toFormattedDateString()}} by <a
+                        href="#">{{$post->user->name}}</a></p>
             <p>
-                <p>{!! $post->content !!}</p>
-                <p><br></p>
+            <p>{!! $post->content !!}</p>
+            <p><br></p>
             </p>
             <div>
                 <a href="/posts/62/zan" type="button" class="btn btn-primary btn-lg">赞</a>
